util: add parse_timestamp() function
Problem: The parse_timestamp() helper function in jobs_table_subcommands
could be used to parse timestamps in other functions in the Python
bindings and would be useful to have in the util.py file.

Move the function out of jobs_table_subcommands.py and into util.py.
